# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Poseidon API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3001",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "access_token": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/area/create": {
      "post": {
        "tags": [
          "Area"
        ],
        "description": "Create a new Area as well as default wallets",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create a new area",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/createAreaModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/pool/create": {
      "post": {
        "tags": [
          "Pool"
        ],
        "description": "Create a new Pool",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create a new pool",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/createPoolModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "$ref": "#/definitions/createPoolResponse"
            }
          }
        }
      }
    },
    "/user/get": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get user information",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ]
      }
    },
    "/user/create-user": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "create sub user ",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create a new sub user",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/createUserRequest"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Register"
        ],
        "description": "Endpoint to sign up user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create a new user",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/SignUpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "$ref": "#/definitions/SignUpRequest"
            }
          }
        }
      }
    },
    "/auth/signupDevUser": {
      "post": {
        "tags": [
          "Register"
        ],
        "description": "Endpoint to sign up user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create a new user",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/SignUpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully obtained.",
            "schema": {
              "$ref": "#/definitions/SignUpRequest"
            }
          }
        }
      }
    },
    "/auth/loginWithUserId": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/token": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/transaction/create": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "description": "Create a new transaction",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create a new transaction",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/createTransactionModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction successfully created.",
            "schema": {
              "$ref": "#/definitions/createTransactionModel"
            }
          }
        }
      }
    },
    "/transaction/get-ranges-by-creator-id": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "description": "Get transactions by creator-id in ranges",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Get transactions by creator-id",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/getTransactionsInRangesByCreatorId"
            }
          }
        ],
        "responses": {}
      }
    },
    "/transaction/get-by-id": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "description": "Get transactions by wallet-id all time",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Get transactions by creator-id",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/getTransactionsInRangesByWalletIdModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/transaction/get-by-ids": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "description": "Get transactions by wallet-id all time",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Get transactions by creator-id",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/getTransactionsInRangesByWalletIdModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/note/create": {
      "post": {
        "tags": [
          "Note"
        ],
        "description": "Create a new note",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create a new note",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/createNoteModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note successfully created.",
            "schema": {
              "$ref": "#/definitions/createNoteModel"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "GenerateTokenRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "SomeSpecialString"
        },
        "pageId": {
          "type": "string",
          "example": "a$b%som3thing"
        }
      },
      "required": [
        "token",
        "pageId"
      ],
      "xml": {
        "name": "GenerateTokenRequest"
      }
    },
    "SignUpRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Nguyen Bang Hoang"
        },
        "email": {
          "type": "string",
          "example": "hoangmrb@gmail.com"
        },
        "phone": {
          "type": "string",
          "example": "0986323066"
        },
        "location": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string",
              "example": "Kim Hoa"
            },
            "ward": {
              "type": "string",
              "example": "Kim Hoa"
            },
            "district": {
              "type": "string",
              "example": "Dong Da"
            },
            "city": {
              "type": "string",
              "example": "Hanoi"
            },
            "province": {
              "type": "string",
              "example": "Hanoi"
            }
          }
        },
        "googleToken": {
          "type": "string",
          "example": "String"
        }
      },
      "xml": {
        "name": "SignUpRequest"
      }
    },
    "createAreaModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Ten Khu Vuc"
        },
        "ownerIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "location": {
          "type": "string",
          "example": "String - location"
        },
        "target": {
          "type": "string",
          "example": "String - Target to be raised!"
        },
        "phone": {
          "type": "string",
          "example": "string"
        }
      },
      "xml": {
        "name": "createAreaModel"
      }
    },
    "createPoolModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "String"
        },
        "areaId": {
          "type": "string",
          "example": "mongoose.SchemaTypes.ObjectId - 5ff6c9d75d8593237c1c5b5d"
        },
        "target": {
          "type": "string",
          "example": "String"
        },
        "unit": {
          "type": "string",
          "example": "String"
        }
      },
      "xml": {
        "name": "createPoolModel"
      }
    },
    "createPoolResponse": {
      "type": "object",
      "properties": {
        "pool": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "String"
            },
            "areaId": {
              "type": "string",
              "example": "mongoose.SchemaTypes.ObjectId - 5ff6c9d75d8593237c1c5b5d"
            },
            "target": {
              "type": "string",
              "example": "String"
            },
            "unit": {
              "type": "string",
              "example": "String"
            }
          }
        },
        "wallet": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "String"
            },
            "ownerId": {
              "type": "string",
              "example": "mongoose.SchemaTypes.ObjectId"
            },
            "unit": {
              "type": "string",
              "example": "String"
            },
            "areaId": {
              "type": "string",
              "example": "mongoose.SchemaTypes.ObjectId"
            },
            "balance": {
              "type": "string",
              "example": "Number"
            },
            "poolId": {
              "type": "string",
              "example": "mongoose.SchemaTypes.ObjectId"
            }
          }
        }
      },
      "xml": {
        "name": "createPoolResponse"
      }
    },
    "getUserModel": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "example": "mongoose.SchemaTypes.ObjectId - 5ff6c9d75d8593237c1c5b5d"
        }
      },
      "xml": {
        "name": "getUserModel"
      }
    },
    "createTransactionModel": {
      "type": "object",
      "properties": {
        "walletId": {
          "type": "string",
          "example": "mongoose.SchemaTypes.ObjectId"
        },
        "creatorId": {
          "type": "string",
          "example": "mongoose.SchemaTypes.ObjectId"
        },
        "createdDate": {
          "type": "number",
          "example": 1610167187000
        },
        "quantity": {
          "type": "number",
          "example": 0
        },
        "title": {
          "type": "string",
          "example": "Transaction's title"
        },
        "customData": {
          "type": "object",
          "properties": {
            "note": {
              "type": "string",
              "example": "Do smt"
            }
          }
        }
      },
      "xml": {
        "name": "createTransactionModel"
      }
    },
    "getTransactionsInRangesByCreatorId": {
      "type": "object",
      "properties": {
        "creatorId": {
          "type": "string",
          "example": "mongoose.SchemaTypes.ObjectId"
        },
        "startDate": {
          "type": "number",
          "example": 1610167187000
        },
        "endDate": {
          "type": "number",
          "example": 1610348546000
        }
      },
      "xml": {
        "name": "getTransactionsInRangesByCreatorId"
      }
    },
    "getTransactionsInRangesByWalletIdModel": {
      "type": "object",
      "properties": {
        "walletId": {
          "type": "string",
          "example": "600c46772a3020039d1ebb0a"
        },
        "startDate": {
          "type": "string",
          "example": "2021-01-24T09:31:27.114Z"
        },
        "endDate": {
          "type": "string",
          "example": "2021-01-28T09:31:27.114Z"
        }
      },
      "xml": {
        "name": "getTransactionsInRangesByWalletIdModel"
      }
    },
    "createNoteModel": {
      "type": "object",
      "properties": {
        "noteType": {
          "type": "string",
          "example": "NoteType"
        },
        "notes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "date": {
                "type": "number",
                "example": 1610167187000
              },
              "content": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "creatorId": {
                "type": "string",
                "example": "mongoose.SchemaType.ObjectId"
              }
            }
          }
        },
        "areaId": {
          "type": "string",
          "example": "mongoose.SchemaType.ObjectId"
        },
        "poolId": {
          "type": "string",
          "example": "mongoose.SchemaType.ObjectId"
        }
      },
      "xml": {
        "name": "createNoteModel"
      }
    },
    "createUserRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "Hoang"
        },
        "password": {
          "type": "string",
          "example": "Bang"
        },
        "email": {
          "type": "string",
          "example": "String"
        },
        "phone": {
          "type": "string",
          "example": "String"
        },
        "location": {
          "type": "object",
          "properties": {
            "city": {
              "type": "string",
              "example": "Ha Noi"
            },
            "district": {
              "type": "string",
              "example": "Dong Da"
            },
            "province": {
              "type": "string",
              "example": "Ha Noi"
            },
            "block": {
              "type": "string",
              "example": "Phuong Lien"
            }
          }
        },
        "verifyEmail": {
          "type": "boolean",
          "example": false
        },
        "avatar": {
          "type": "string",
          "example": "string"
        },
        "detailInformation": {
          "type": "object",
          "properties": {
            "isRoot": {
              "type": "boolean",
              "example": false
            },
            "subAccounts": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "xml": {
        "name": "createUserRequest"
      }
    },
    "findWalletsByIdRequest": {
      "type": "object",
      "properties": {
        "areaId": {
          "type": "string",
          "example": "String"
        }
      },
      "xml": {
        "name": "findWalletsByIdRequest"
      }
    }
  }
}